requirements.txt
openai==1.59.8
streamlit==1.41.1
a2a-sdk
httpx==0.27.0
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.0